Moved views into exchange.py
